requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.30.0
peft>=0.4.0
yt-dlp>=2023.7.6
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
datasets>=2.12.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0
gradio
ultralytics
git+https://github.com/openai/CLIP.git